Add search bar, file uploader, file table, and file row components for knowledgebase
